% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{binChi}
\alias{binChi}
\alias{binMI}
\alias{binAbsDif}
\title{Statistics for bins}
\usage{
binChi(bins, agg = sum)

binMI(bins, agg = sum)

binAbsDif(bins, agg = sum)
}
\arguments{
\item{bins}{a list of bins, each a list with elements `x`, `y`,
`depth`, `bnds` (list with elements `x` and `y`), `expn`, `n`}

\item{agg}{function which is aggregates the individual statistics
computed over each bin}
}
\value{
A list with elements `residuals` and `stat` reporting the
individual statistic values (possibly transformed) and the
aggegrated statistic value.
}
\description{
These functions compute statistics based on observed
and expected counts for a list of bins.
}
\details{
Binstatistics

Three functions are provided by default, `binChi`
computes the chi-squared statistic by taking the squared
difference between observed and expected counts and dividing this
by the expected counts. `binMi` computes the mutual information
for each bin using the observed and expected counts. Finally,
`binAbsDif` computes the absolute difference between observed
and expected counts. Each function first computes a value on
every bin independently and stores all these values in memory
before using the function provided in the optional argument `agg`
to aggregate these values.
}
\section{Functions}{
\itemize{
\item \code{binChi()}: Chi-squared statistic

\item \code{binMI()}: Mutual information

\item \code{binAbsDif()}: Absolute difference between observed
and expected

}}
\author{
Chris Salahub
}
