% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitters.R
\name{halfCutTie}
\alias{halfCutTie}
\title{Halve continuously to break ties}
\usage{
halfCutTie(bin, xscore, yscore)
}
\arguments{
\item{bin}{a bin to be split with elements `x`, `y`, `depth`,
`bnds` (list with elements `x` and `y`), `expn`, `n`}

\item{xscore}{numeric value giving the score for all splits along
x}

\item{yscore}{numeric value giving the score for all splits along
y}
}
\value{
A list of two bins resulting from the split of `bin` in
half along the margin corresponding to the larger score.
}
\description{
This function halves a bin based on the midpoint of
the bounds along whichever margin produces the larger score.
}
\details{
The goal of this function is to break ties within bin
splitting in a way which prevents very small or lopsided bins from
forming, a common problem with the `halfSplit` function
}
\author{
Chris Salahub
}
